Fix hooks implementation issues
1. Combine reducers used in `useChangeSelect` and `useSourceExpansion` to a single `useCollection` hook.
2. Move `useCustomEqualIdentifier` as a base helper hook.
3. Fix an issue where `useSourceExpansion` incorrectly use `oldSource` input.
