+++ src/hooks/useChangeSelect.js
@@ -1,43 +1,21 @@
-import {useReducer, useEffect} from 'react';
+import {useEffect} from 'react';
 import {getChangeKey} from '../utils';
-
-const useCollection = multiple => {
-    const update = (collection, {type, value}) => {
-        switch (type) {
-            case 'clear':
-                return collection.length ? [] : collection;
-            case 'toggle': {
-                if (multiple) {
-                    return collection.includes(value)
-                        ? collection.filter(item => item !== value)
-                        : collection.concat(value);
-                }
-
-                return [value];
-            }
-            default:
-                return collection;
-        }
-    };
-
-    const [collection, dispatch] = useReducer(update, []);
-    return {
-        collection,
-        clear() {
-            dispatch({type: 'clear'});
-        },
-        toggle(value) {
-            dispatch({value, type: 'toggle'});
-        },
-    };
-};
+import {useCollection} from './helpers';
 
 export default (hunks, {multiple = false} = {}) => {
-    const {collection, clear, toggle} = useCollection(multiple);
+    const {collection, clear, toggle, only} = useCollection();
     useEffect(clear, [hunks]);
 
     return [
         collection,
-        ({change}) => toggle(getChangeKey(change)),
+        ({change}) => {
+            const changeKey = getChangeKey(change);
+            if (multiple) {
+                toggle(changeKey);
+            }
+            else {
+                only(changeKey);
+            }
+        },
     ];
 };
