+++ src/hooks/useSourceExpansion.js
@@ -1,37 +1,14 @@
-import {useReducer, useEffect, useMemo} from 'react';
+import {useEffect, useMemo} from 'react';
 import {expandFromRawCode} from '../utils';
-
-const useAppendableCollection = () => {
-    const update = (collection, {type, value}) => {
-        switch (type) {
-            case 'clear':
-                return collection.length ? [] : collection;
-            case 'push':
-                return [...collection, value];
-            default:
-                return collection;
-        }
-    };
-
-    const [collection, dispatch] = useReducer(update, []);
-    return {
-        collection,
-        clear() {
-            dispatch({type: 'clear'});
-        },
-        push(value) {
-            dispatch({value, type: 'push'});
-        },
-    };
-};
+import {useCollection} from './helpers';
 
 export default (hunks, oldSource) => {
-    const {collection: expandedRanges, clear, push} = useAppendableCollection();
+    const {collection: expandedRanges, clear, push} = useCollection();
     useEffect(clear, [hunks, oldSource]);
     const linesOfOldSource = useMemo(() => (oldSource || '').split('\n'), [oldSource]);
     const renderingHunks = useMemo(
         () => {
-            if (!oldSource) {
+            if (!linesOfOldSource.length) {
                 return hunks;
             }
